Move contents of libstd_unicode into libcore
