+++ src/libcore/unicode/bool_trie.rs
