+++ src/libcore/unicode/char.rs
@@ -28,30 +28,30 @@
 
 #![stable(feature = "rust1", since = "1.0.0")]
 
-use core::char::CharExt as C;
-use core::iter::FusedIterator;
-use core::fmt::{self, Write};
-use tables::{conversions, derived_property, general_category, property};
+use char::CharExt as C;
+use iter::FusedIterator;
+use fmt::{self, Write};
+use unicode::tables::{conversions, derived_property, general_category, property};
 
 // stable re-exports
 #[stable(feature = "rust1", since = "1.0.0")]
-pub use core::char::{MAX, from_digit, from_u32, from_u32_unchecked};
+pub use char::{MAX, from_digit, from_u32, from_u32_unchecked};
 #[stable(feature = "rust1", since = "1.0.0")]
-pub use core::char::{EscapeDebug, EscapeDefault, EscapeUnicode};
+pub use char::{EscapeDebug, EscapeDefault, EscapeUnicode};
 #[stable(feature = "decode_utf16", since = "1.9.0")]
-pub use core::char::REPLACEMENT_CHARACTER;
+pub use char::REPLACEMENT_CHARACTER;
 #[stable(feature = "char_from_str", since = "1.20.0")]
-pub use core::char::ParseCharError;
+pub use char::ParseCharError;
 
 // unstable re-exports
 #[stable(feature = "try_from", since = "1.26.0")]
-pub use core::char::CharTryFromError;
+pub use char::CharTryFromError;
 #[unstable(feature = "decode_utf8", issue = "33906")]
-pub use core::char::{DecodeUtf8, decode_utf8};
+pub use char::{DecodeUtf8, decode_utf8};
 #[unstable(feature = "unicode", issue = "27783")]
-pub use tables::{UNICODE_VERSION};
+pub use unicode::tables::{UNICODE_VERSION};
 #[unstable(feature = "unicode", issue = "27783")]
-pub use version::UnicodeVersion;
+pub use unicode::version::UnicodeVersion;
 
 /// Returns an iterator that yields the lowercase equivalent of a `char`.
 ///
